-typo fixed

+++ parser/parse_content.c
@@ -45,9 +45,9 @@
 
 
 #define is_mime_char(_c_) \
-	(isalpha(_c_) || (_c_)=='-' || (_c_)=='+')
+	(isalpha((int)_c_) || (_c_)=='-' || (_c_)=='+')
 #define is_char_equal(_c_,_cs_) \
-	( (isalpha(_c_)?(((_c_)|0x20)==(_cs_)):((_c_)==(_cs_)))==1 )
+	( (isalpha((int)_c_)?(((_c_)|0x20)==(_cs_)):((_c_)==(_cs_)))==1 )
 
 
 /*
